Rename test packages to wasm as well
