+++ app/wasm/helpers_test.go
@@ -1,4 +1,4 @@
-package wasmtest
+package wasm
 
 import (
 	"testing"
